fix: voice message insert — pass duration_seconds as int, not long long
Drogon sends long long as 8-byte BIGINT in binary protocol, but the
duration_seconds column is INTEGER (4 bytes). PostgreSQL rejects the
bind parameter with "incorrect binary data format in bind parameter 6".

Remove the (long long) cast so durationSecs is passed as its declared
int type, matching the column's INTEGER format.

+++ backend/src/controllers/MessagesController.cpp
@@ -244,7 +244,7 @@ void MessagesController::sendMessage(const drogon::HttpRequestPtr& req,
                                          chatId, me, content, msgType, resolvedStickerId, replyToMsgId);
                     else if (resolvedFileId > 0 && durationSecs > 0)
                         db->execSqlAsync(sql, std::move(onInserted), std::move(onErr),
-                                         chatId, me, content, msgType, resolvedFileId, (long long)durationSecs, replyToMsgId);
+                                         chatId, me, content, msgType, resolvedFileId, durationSecs, replyToMsgId);
                     else if (resolvedFileId > 0)
                         db->execSqlAsync(sql, std::move(onInserted), std::move(onErr),
                                          chatId, me, content, msgType, resolvedFileId, replyToMsgId);
